Brandon dev (#5)
* US04 test case implemented

* Implemented US04 test case based on US03 test case code design
Made one small change to US03 and the main function. To lower the number of lines of code in the main function and only add in the very bare minimum necessary to call out to the test functions, a super function was created in both US03 and US04 called Execute. This function just calls the module's Test and Display_Results functions and allows the region in the main function, where the calling of the tests occur, to be much cleaner and easier to read.